Extract titan-specific functions into new file
